create part which define first move,server,imperative

+++ src/main/scala/game/Player.scala
@@ -3,6 +3,6 @@ package game
 /**
  * Created by aleapv on 16.08.2015.
  */
-case class Player(number: Short) {
+case class Player(number: Integer) {
 
 }
